refactor: finish moving code to dart_pipeline.metrics
