+++ src/dart_pipeline/collate.py
@@ -36,18 +36,8 @@ This will create a `data/sources/meteorological/aphrodite-daily-mean-temp` folde
 into which data will be downloaded.
 """
 
-import logging
-import base64
-import re
-from datetime import date, timedelta
-from typing import Final, Callable
-
-from bs4 import BeautifulSoup
-import requests
-
-from .constants import TERRACLIMATE_METRICS, PERU_REGIONS
-from .types import URLCollection, DataFile, PartialDate
-from .util import daterange, use_range, get_country_name
+from .types import URLCollection
+from .util import get_country_name
 
 
 def gadm_data(iso3: str) -> URLCollection:
@@ -68,284 +58,6 @@ def gadm_data(iso3: str) -> URLCollection:
         ],
         relative_path=iso3,
     )
-
-
-def relative_wealth_index(iso3: str) -> URLCollection:
-    """This dataset contains the relative wealth index, which is the relative
-    standard of living, obtained from connectivity data, satellite imagery and
-    other sources. Cite the following if using this dataset:
-
-        Microestimates of wealth for all low- and middle-income countries.
-        Guanghua Chi, Han Fang, Sourav Chatterjee, Joshua E. Blumenstock
-        Proceedings of the National Academy of Sciences
-        Jan 2022, 119 (3) e2113658119; DOI: 10.1073/pnas.2113658119
-
-    Upstream URL: https://data.humdata.org/dataset/relative-wealth-index
-    """
-    # Validate input parameter
-    if not iso3:
-        raise ValueError("No ISO3 code has been provided")
-
-    # Search the webpage for the link(s) to the dataset(s)
-    url = "https://data.humdata.org/dataset/relative-wealth-index"
-    if (r := requests.get(url)).status_code == 200:
-        # Search for a URL in the HTML content
-        soup = BeautifulSoup(r.text, "html.parser")
-        # Find all anchor tags (<a>) with href attribute containing the ISO3
-        target = iso3.lower()
-        links = soup.find_all("a", href=lambda href: href and target in href)  # type: ignore
-        # Return the first link found
-        if links:
-            csvs = [link["href"] for link in links if "csv" in link["href"]]
-            return URLCollection("https://data.humdata.org", csvs)
-        else:
-            raise ValueError(f'Could not find a link containing "{target}"')
-    else:
-        raise ValueError(f'Bad response for page: "{r.status_code}"')
-
-
-def ministerio_de_salud_peru_data() -> list[DataFile]:
-    """
-    Download data from the Ministerio de Salud (Peru).
-
-    https://www.dge.gob.pe/sala-situacional-dengue
-    """
-    pages = ["Nacional_dengue"] + ["sala_dengue_" + region for region in PERU_REGIONS]
-    # If the user specifies that only one dataset should be downloaded
-    data: list[DataFile] = []
-    for page in pages:
-        url = "https://www.dge.gob.pe/sala-situacional-dengue/uploads/" + f"{page}.html"
-        print(f"Accessing {url}")
-        response = requests.get(url)
-        # Raise an exception for bad response status
-        response.raise_for_status()
-        # Parse HTML content
-        soup = BeautifulSoup(response.content, "html.parser")
-        # Find links with the onclick attribute in both <a> and <button> tags
-        onclick_elements = soup.find_all(
-            lambda tag: tag.name in ["a", "button"] and tag.has_attr("onclick")
-        )
-        links = [element.get("onclick") for element in onclick_elements]
-        if not links:
-            raise ValueError("No links found on the page")
-
-        for link in links:
-            # Search the link for the data embedded in it
-            matches = re.findall(r"base64,(.*?)(?='\).then)", link, re.DOTALL)
-            if matches:
-                base64_string = matches[0]
-            else:
-                raise ValueError("No data found embedded in the link")
-
-            # Search the link for the filename
-            matches = re.findall(r"a\.download = '(.*?)';\s*a\.click", link)
-            if matches:
-                # There is an actual filename for this data
-                filename = matches[0]
-            else:
-                # Use the page name for the file
-                filename = page + ".xlsx"
-
-            file = DataFile(filename, ".", base64.b64decode(base64_string))
-            data.append(file)
-    return data
-
-
-def aphrodite_precipitation_data() -> list[URLCollection]:
-    "APHRODITE Daily accumulated precipitation (V1901) [requires account]"
-    base_url = "http://aphrodite.st.hirosaki-u.ac.jp"
-    return [
-        # 0.05 degree
-        URLCollection(
-            f"{base_url}/product/APHRO_V1901/APHRO_MA/005deg",
-            ["APHRO_MA_PREC_CLM_005deg_V1901.ctl.gz"],
-        ),
-        # 0.25 degree
-        URLCollection(
-            f"{base_url}/product/APHRO_V1901/APHRO_MA/025deg",
-            [
-                "APHRO_MA_025deg_V1901.2015.gz",
-                "APHRO_MA_025deg_V1901.ctl.gz",
-            ],
-        ),
-        # 0.25 degree nc
-        URLCollection(
-            f"{base_url}/product/APHRO_V1901/APHRO_MA/025deg_nc",
-            ["APHRO_MA_025deg_V1901.2015.nc.gz"],
-        ),
-        # 0.50 degree
-        URLCollection(
-            f"{base_url}/product/APHRO_V1901/APHRO_MA/050deg",
-            [
-                "APHRO_MA_050deg_V1901.2015.gz",
-                "APHRO_MA_050deg_V1901.ctl.gz",
-            ],
-        ),
-        # 0.50 degree nc
-        URLCollection(
-            f"{base_url}/product/APHRO_V1901/APHRO_MA/050deg_nc",
-            ["APHRO_MA_050deg_V1901.2015.nc.gz"],
-        ),
-    ]
-
-
-def aphrodite_temperature_data() -> list[URLCollection]:
-    "APHRODITE Daily mean temperature product (V1808) [requires account]"
-
-    base_url = "http://aphrodite.st.hirosaki-u.ac.jp"
-    return [
-        URLCollection(
-            f"{base_url}/product/APHRO_V1808_TEMP/APHRO_MA/050deg_nc",
-            [
-                "APHRO_MA_TAVE_050deg_V1808.2015.nc.gz",  # 19 MB
-                "APHRO_MA_TAVE_050deg_V1808.nc.ctl.gz",  # 347 B
-            ],
-        ),
-        URLCollection(
-            f"{base_url}/product/APHRO_V1808_TEMP/APHRO_MA/050deg",
-            ["read_aphro_v1808.f90"],
-        ),
-        URLCollection(
-            f"{base_url}/product/APHRO_V1808_TEMP/APHRO_MA/025deg_nc",
-            [
-                "APHRO_MA_TAVE_025deg_V1808.2015.nc.gz",  # 64 MB
-                "APHRO_MA_TAVE_025deg_V1808.nc.ctl.gz",  # 485 B
-            ],
-        ),
-        URLCollection(
-            f"{base_url}/product/APHRO_V1808_TEMP/APHRO_MA/025deg",
-            [
-                "APHRO_MA_TAVE_025deg_V1808.2015.gz",  # 64 MB
-                "APHRO_MA_TAVE_025deg_V1808.ctl.gz",  # 312 B
-                "read_aphro_v1808.f90",  # 2.6 KB
-            ],
-        ),
-        URLCollection(
-            f"{base_url}/product/APHRO_V1808_TEMP/APHRO_MA/005deg_nc",
-            [
-                "APHRO_MA_TAVE_CLM_005deg_V1808.nc.gz",  # 1.2 GB
-            ],
-        ),
-        URLCollection(
-            f"{base_url}/product/APHRO_V1808_TEMP/APHRO_MA/005deg",
-            [
-                "APHRO_MA_TAVE_CLM_005deg_V1808.ctl.gz",  # 334 B
-                "APHRO_MA_TAVE_CLM_005deg_V1808.grd.gz",  # 1.4 GB
-                "read_aphro_clm_v1808.f90",  # 2.1 KB
-            ],
-        ),
-    ]
-
-
-def chirps_rainfall_data(partial_date: str) -> list[URLCollection]:
-    """
-    CHIRPS Rainfall Estimates from Rain Gauge, Satellite Observations.
-
-    "CHIRPS" stands for Climate Hazards Group InfraRed Precipitation with
-    Station.
-
-    Data is in TIF format (.tif.gz), not COG format (.cog).
-    """
-    pdate = PartialDate.from_string(partial_date)
-    base_url = "https://data.chc.ucsb.edu"
-    fmt = "tifs"  # cogs is unsupported at the moment
-    chirps_first_year: Final[int] = 1981
-    chirps_first_month: Final[date] = date(1981, 1, 1)
-    urls: list[URLCollection] = []
-
-    if pdate.month:
-        use_range(pdate.month, 1, 12, "Month range")
-
-    today = date.today()
-    use_range(pdate.year, chirps_first_year, today.year, "CHIRPS annual data range")
-    urls.append(
-        URLCollection(
-            f"{base_url}/products/CHIRPS-2.0/global_annual/{fmt}",
-            [f"chirps-v2.0.{pdate.year}.tif"],
-            relative_path="global_annual",
-        )
-    )
-
-    if pdate.month:
-        # Download the monthly data for the year and month provided
-        month_requested = date(pdate.year, pdate.month, 1)
-        this_month = date(today.year, today.month, 1)
-        if chirps_first_month <= month_requested < this_month:
-            base = f"{base_url}/products/CHIRPS-2.0/global_monthly/{fmt}"
-            files = [f"chirps-v2.0.{pdate.year}.{pdate.month:02d}.tif.gz"]
-            path = f"global_monthly/{pdate.year}"
-            urls.append(URLCollection(base, files, relative_path=path))
-        else:
-            logging.warning(
-                "Monthly data is only available from "
-                + f"{chirps_first_year}-01 onwards"
-            )
-            return urls
-
-        # Download the daily data for the year and month provided
-        end = date(int(pdate.year), int(pdate.month) + 1, 1)
-        end = end - timedelta(days=1)
-        base = (
-            f"{base_url}/products/CHIRPS-2.0/global_daily/" + f"{fmt}/p05/{pdate.year}"
-        )
-        files = [
-            f"chirps-v2.0.{str(day).replace('-', '.')}.tif.gz"
-            for day in daterange(month_requested, end)
-        ]
-        path = f"global_daily/{pdate.year}/{pdate.month:02d}"
-        urls.append(URLCollection(base, files, relative_path=path))
-
-    return urls
-
-
-def terraclimate_data(year: int) -> URLCollection:
-    """TerraClimate gridded temperature, precipitation data.
-
-    TerraClimate is a dataset of monthly climate and climatic water balance for
-    terrestrial surfaces from 1958--2023. Data have a monthly temporal
-    resolution and 4 km (1/24th degree) spatial resolution.
-
-    Upstream URL: https://www.climatologylab.org/terraclimate.html
-    """
-    use_range(year, 1958, 2023, "Terraclimate year range")
-    return URLCollection(
-        "https://climate.northwestknowledge.net/TERRACLIMATE-DATA",
-        # 2023, capitalisation of PDSI changed
-        [f"TerraClimate_PDSI_{year}.nc"]
-        + [f"TerraClimate_{metric}_{year}.nc" for metric in TERRACLIMATE_METRICS],
-    )
-
-
-def meta_pop_density_data(iso3: str) -> URLCollection:
-    """
-    Download Population Density Maps from Data for Good at Meta.
-
-    Documentation:
-    https://dataforgood.facebook.com/dfg/docs/high-resolution-population-density-maps-demographic-estimates-documentation
-    """
-    country = get_country_name(iso3)
-    print(f"Country:   {country}")
-    # Main webpage
-    url = (
-        "https://data.humdata.org/dataset/"
-        f"{country.lower().replace(' ', '-')}-high-resolution-population-"
-        "density-maps-demographic-estimates"
-    )
-    if (response := requests.get(url)).status_code == 200:
-        # Search for a URL in the HTML content
-        soup = BeautifulSoup(response.text, "html.parser")
-        # Find all anchor tags (<a>) with href attribute containing the ISO3
-        target = iso3.lower()
-        if links := soup.find_all("a", href=lambda href: href and target in href):  # type: ignore
-            return URLCollection(
-                "https://data.humdata.org",
-                [link["href"] for link in links if link["href"].endswith(".zip")],
-                relative_path=iso3,
-            )
-        else:
-            raise ValueError(f'Could not find a link containing "{target}"')
-    else:
-        raise ValueError(f'Bad response for page: "{response.status_code}"')
 
 
 def worldpop_pop_count_data(iso3: str) -> URLCollection:
@@ -391,24 +103,3 @@ def worldpop_pop_density_data(iso3: str) -> URLCollection:
         ],
         relative_path=iso3,
     )
-
-
-REQUIRES_AUTH = [
-    "meteorological/aphrodite-daily-precip",
-    "meteorological/aphrodite-daily-mean-temp",
-]
-
-SOURCES: dict[
-    str, Callable[..., URLCollection | list[URLCollection] | list[DataFile]]
-] = {
-    "economic/relative-wealth-index": relative_wealth_index,
-    "epidemiological/dengue/peru": ministerio_de_salud_peru_data,
-    "geospatial/gadm": gadm_data,
-    "meteorological/aphrodite-daily-mean-temp": aphrodite_temperature_data,
-    "meteorological/aphrodite-daily-precip": aphrodite_precipitation_data,
-    "meteorological/chirps-rainfall": chirps_rainfall_data,
-    "meteorological/terraclimate": terraclimate_data,
-    "sociodemographic/meta-pop-density": meta_pop_density_data,
-    "sociodemographic/worldpop-count": worldpop_pop_count_data,
-    "sociodemographic/worldpop-density": worldpop_pop_density_data,
-}
