+++ src/dart_pipeline/constants.py
@@ -22,48 +22,6 @@ MSG_PROCESS = COL_CYAN + " process" + COL_OFF
 # Smallest single-precision floating-point number
 MIN_FLOAT = -3.4028234663852886e38
 
-TERRACLIMATE_METRICS = [
-    "aet",  # water_evaporation_amount_mm
-    "def",  # water_potential_evaporation_amount_minus_water_evaporatio
-    "pdsi",  # palmer_drought_severity_index (unitless)
-    "pet",  # water_potential_evaporation_amount_mm
-    "ppt",  # precipitation_amount_mm
-    "q",  # runoff_amount_mm
-    "soil",  # soil_moisture_content_mm
-    "srad",  # downwelling_shortwave_flux_in_air_W_per_m_squared
-    "swe",  # liquid_water_content_of_surface_snow_mm
-    "tmax",  # air_temperature_max_degC
-    "tmin",  # air_temperature_min_degC
-    "vap",  # water_vapor_partial_pressure_in_air_kPa
-    "vpd",  # vapor_pressure_deficit_kPa
-    "ws",  # wind_speed_m_per_s
-]
-
-PERU_REGIONS = [
-    "AMAZONAS",
-    "ANCASH",
-    "AREQUIPA",
-    "AYACUCHO",
-    "CAJAMARCA",
-    "CALLAO",
-    "CUSCO",
-    "HUANUCO",
-    "ICA",
-    "JUNIN",
-    "LA LIBERTAD",
-    "LAMBAYEQUE",
-    "LIMA",
-    "LORETO",
-    "MADRE DE DIOS",
-    "MOQUEGUA",
-    "PASCO",
-    "PIURA",
-    "PUNO",
-    "SAN MARTIN",
-    "TUMBES",
-    "UCAYALI",
-]
-
 # Column names in the output CSVs
 OUTPUT_COLUMNS = [
     "iso3",
