+++ tests/metrics/test_meta_relative_wealth_index.py
@@ -1,17 +1,71 @@
 """Tests for functions in population-weighted/relative-wealth-index.py."""
 
 from unittest import mock
+from pathlib import Path
 
 import geopandas as gpd
 import pandas as pd
 import shapely.geometry
+import requests_mock
 
+from dart_pipeline.types import URLCollection
 from dart_pipeline.metrics.meta_relative_wealth_index import (
     get_admin_region,
     process_gadm_rwi,
     process_rwi_point_estimate,
     process_gadm_popdensity_rwi,
+    fetch_relative_wealth_index,
+    meta_pop_density_data,
 )
+
+
+def test_meta_pop_density_data():
+    web_snapshot = Path(
+        "tests/webarchive/vietnam-high-resolution-population-density-maps-demographic-estimates.html"
+    )
+    url = "https://data.humdata.org/dataset/vietnam-high-resolution-population-density-maps-demographic-estimates"
+    dataset = "/dataset/191b04c5-3dc7-4c2a-8e00-9c0bdfdfbf9d/resource"
+    with requests_mock.Mocker() as m:
+        m.get(
+            url,
+            text=web_snapshot.read_text(),
+        )
+        assert meta_pop_density_data("VNM") == URLCollection(
+            "https://data.humdata.org",
+            files=[
+                f"{dataset}/b60dab07-0a27-47f3-894d-02e9fbee6473/download/vnm_children_under_five_2020_csv.zip",
+                f"{dataset}/a2ca47a3-b7d7-4a39-a605-161d4790c6f9/download/vnm_children_under_five_2020_geotiff.zip",
+                f"{dataset}/77edfcc3-d037-4233-9b16-7ebb684b4752/download/vnm_elderly_60_plus_2020_csv.zip",
+                f"{dataset}/61b6a396-ffcc-4710-a453-174914822164/download/vnm_elderly_60_plus_2020_geotiff.zip",
+                f"{dataset}/e0d42fbb-2436-4f3e-afd1-83d1ac6314b2/download/vnm_men_2020_csv.zip",
+                f"{dataset}/76990cd9-a2dc-44e6-809c-492698b090c0/download/vnm_men_2020_geotiff.zip",
+                f"{dataset}/0bd525fa-3f63-447b-9afa-3d5075657ce4/download/vnm_women_2020_csv.zip",
+                f"{dataset}/5bd99468-ea44-4a50-b64b-54343b182842/download/vnm_women_2020_geotiff.zip",
+                f"{dataset}/5727eb4e-ce2a-4250-a9a0-2603d889ff02/download/vnm_women_of_reproductive_age_15_49_2020_csv.zip",
+                f"{dataset}/855e27d7-d191-440d-936a-ecedfded238d/download/vnm_women_of_reproductive_age_15_49_2020_geotiff.zip",
+                f"{dataset}/aa6dd9dd-eecd-4f3d-b6b2-fa0f0b5faa70/download/vnm_youth_15_24_2020_csv.zip",
+                f"{dataset}/e36dce9f-ea1d-4933-8a37-dc5838df11f0/download/vnm_youth_15_24_2020_geotiff.zip",
+                f"{dataset}/fade8620-0935-4d26-b0c6-15515dd4bf8b/download/vnm_general_2020_geotiff.zip",
+                f"{dataset}/0fbf4055-7091-4041-a7ea-25f057debd7c/download/vnm_general_2020_csv.zip",
+            ],
+            relative_path="VNM",
+        )
+
+
+def test_relative_wealth_index():
+    web_snapshot = Path("tests/webarchive/relative-wealth-index.html")
+    base_url = "https://data.humdata.org/dataset/relative-wealth-index"
+    with requests_mock.Mocker() as m:
+        m.get(
+            base_url,
+            text=web_snapshot.read_text(),
+        )
+        assert fetch_relative_wealth_index("VNM") == URLCollection(
+            "https://data.humdata.org",
+            [
+                "/dataset/76f2a2ea-ba50-40f5-b79c-db95d668b843/resource/06d29bc0-5a4c-4be0-be1a-c546a9be540c/download/vnm_relative_wealth_index.csv"
+            ],
+        )
 
 
 def test_get_admin_region():
@@ -56,10 +110,10 @@ def test_process_gadm_rwi(mock_plot, mock_get_country, mock_get_path):
     with (
         mock.patch("geopandas.read_file", return_value=gdf),
         mock.patch("pandas.read_csv", return_value=rwi),
-        mock.patch(
-            "pandas.DataFrame.parallel_apply",
-            side_effect=lambda func, axis: rwi.apply(func, axis=axis),
-        ),
+        # mock.patch(
+        #     "pandas.DataFrame.parallel_apply",
+        #     side_effect=lambda func, axis: rwi.apply(func, axis=axis),
+        # ),
     ):
         output_rwi = process_gadm_rwi("VNM", 1, plots=False)
